# feat: reduce memory allocs to optimize performance