--- a/lexerStream.go
+++ b/lexerStream.go
@@ -1,28 +1,35 @@
 package govaluate
 
+import "sync"
+
 type lexerStream struct {
-	source   []rune
-	position int
-	length   int
+	sourceString string
+	source       []rune
+	position     int
+	length       int
+}
+
+var lexerStreamPool = sync.Pool{
+	New: func() interface{} {
+		return new(lexerStream)
+	},
 }
 
 func newLexerStream(source string) *lexerStream {
-
-	var ret *lexerStream
-	var runes []rune
-
-	for _, character := range source {
-		runes = append(runes, character)
+	ret := lexerStreamPool.Get().(*lexerStream)
+	if ret.source == nil {
+		ret.source = make([]rune, 0, len(source))
 	}
-
-	ret = new(lexerStream)
-	ret.source = runes
-	ret.length = len(runes)
+	for _, character := range source {
+		ret.source = append(ret.source, character)
+	}
+	ret.sourceString = source
+	ret.position = 0
+	ret.length = len(ret.source)
 	return ret
 }
 
 func (this *lexerStream) readCharacter() rune {
-
 	character := this.source[this.position]
 	this.position += 1
 	return character
@@ -34,4 +41,9 @@ func (this *lexerStream) rewind(amount int) {
 
 func (this lexerStream) canRead() bool {
 	return this.position < this.length
+}
+
+func (this *lexerStream) close() {
+	this.source = this.source[:0]
+	lexerStreamPool.Put(this)
 }