--- a/parsing.go
+++ b/parsing.go
@@ -7,13 +7,19 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 	"unicode"
+	"unicode/utf8"
+)
+
+var (
+	averageTokens = 1
+	samples       = make([]int, 0, 10)
 )
 
 func parseTokens(expression string, functions map[string]ExpressionFunction) ([]ExpressionToken, error) {
-
-	var ret []ExpressionToken
+	ret := make([]ExpressionToken, 0, averageTokens)
 	var token ExpressionToken
 	var stream *lexerStream
 	var state lexerState
@@ -43,6 +49,18 @@ func parseTokens(expression string, functions map[string]ExpressionFunction) ([]
 		// append this valid token
 		ret = append(ret, token)
 	}
+	stream.close()
+	if len(samples) == cap(samples) {
+		copy(samples, samples[1:])
+		samples[len(samples)-1] = len(ret)
+	} else {
+		samples = append(samples, len(ret))
+	}
+	total := 0
+	for _, val := range samples {
+		total += val
+	}
+	averageTokens = total / len(samples)
 
 	err = checkBalance(ret)
 	if err != nil {
@@ -140,30 +158,23 @@ func readToken(stream *lexerStream, state lexerState, functions map[string]Expre
 		if unicode.IsLetter(character) {
 
 			tokenString = readTokenUntilFalse(stream, isVariableName)
-
-			tokenValue = tokenString
-			kind = VARIABLE
-
-			// boolean?
-			if tokenValue == "true" {
-
+			switch tokenString {
+			case "true":
 				kind = BOOLEAN
 				tokenValue = true
-			} else {
-
-				if tokenValue == "false" {
-
-					kind = BOOLEAN
-					tokenValue = false
-				}
-			}
-
-			// textual operator?
-			if tokenValue == "in" || tokenValue == "IN" {
-
+			case "false":
+				kind = BOOLEAN
+				tokenValue = false
+			case "in":
+				fallthrough
+			case "IN":
 				// force lower case for consistency
 				tokenValue = "in"
 				kind = COMPARATOR
+			default:
+				// This causes an alloc, avoid it if we can
+				tokenValue = tokenString
+				kind = VARIABLE
 			}
 
 			// function?
@@ -284,37 +295,50 @@ func readTokenUntilFalse(stream *lexerStream, condition func(rune) bool) string 
 	return ret
 }
 
+var tokenBufferPool = sync.Pool{
+	New: func() interface{} {
+		return &bytes.Buffer{}
+	},
+}
+
 /*
 Returns the string that was read until the given [condition] was false, or whitespace was broken.
 Returns false if the stream ended before whitespace was broken or condition was met.
 */
 func readUntilFalse(stream *lexerStream, includeWhitespace bool, breakWhitespace bool, allowEscaping bool, condition func(rune) bool) (string, bool) {
 
-	var tokenBuffer bytes.Buffer
+	tokenBuffer := tokenBufferPool.Get().(*bytes.Buffer)
 	var character rune
-	var conditioned bool
 
-	conditioned = false
+	startPosition := stream.position
+	reuseString := true
+	trimString := false
+	conditioned := false
 
 	for stream.canRead() {
 
 		character = stream.readCharacter()
+		if character > utf8.RuneSelf {
+			// International runes, we can't just grab from the string in this case
+			reuseString = false
+		}
 
 		// Use backslashes to escape anything
 		if allowEscaping && character == '\\' {
-
+			reuseString = false
 			character = stream.readCharacter()
 			tokenBuffer.WriteString(string(character))
 			continue
 		}
 
 		if unicode.IsSpace(character) {
-
 			if breakWhitespace && tokenBuffer.Len() > 0 {
 				conditioned = true
+				trimString = true
 				break
 			}
 			if !includeWhitespace {
+				reuseString = false
 				continue
 			}
 		}
@@ -328,7 +352,21 @@ func readUntilFalse(stream *lexerStream, includeWhitespace bool, breakWhitespace
 		}
 	}
 
-	return tokenBuffer.String(), conditioned
+	// This reduces allocations by just reusing parts of the original source string if applicable
+	if reuseString {
+		tokenBuffer.Reset()
+		tokenBufferPool.Put(tokenBuffer)
+		ret := stream.sourceString[startPosition:stream.position]
+		if trimString {
+			ret = ret[:len(ret)-1]
+		}
+		return ret, conditioned
+	}
+
+	ret := tokenBuffer.String()
+	tokenBuffer.Reset()
+	tokenBufferPool.Put(tokenBuffer)
+	return ret, conditioned
 }
 
 /*
@@ -395,6 +433,8 @@ func checkBalance(tokens []ExpressionToken) error {
 		}
 	}
 
+	stream.close()
+
 	if parens != 0 {
 		return errors.New("Unbalanced parenthesis")
 	}
@@ -448,6 +488,12 @@ func isNotClosingBracket(character rune) bool {
 	return character != ']'
 }
 
+type timeFormat struct {
+	format    string
+	minLength int
+	maxLength int
+}
+
 /*
 Attempts to parse the [candidate] as a Time.
 Tries a series of standardized date formats, returns the Time if one applies,
@@ -458,26 +504,34 @@ func tryParseTime(candidate string) (time.Time, bool) {
 	var ret time.Time
 	var found bool
 
-	timeFormats := [...]string{
-		time.ANSIC,
-		time.UnixDate,
-		time.RubyDate,
-		time.Kitchen,
-		time.RFC3339,
-		time.RFC3339Nano,
-		"2006-01-02",                         // RFC 3339
-		"2006-01-02 15:04",                   // RFC 3339 with minutes
-		"2006-01-02 15:04:05",                // RFC 3339 with seconds
-		"2006-01-02 15:04:05-07:00",          // RFC 3339 with seconds and timezone
-		"2006-01-02T15Z0700",                 // ISO8601 with hour
-		"2006-01-02T15:04Z0700",              // ISO8601 with minutes
-		"2006-01-02T15:04:05Z0700",           // ISO8601 with seconds
-		"2006-01-02T15:04:05.999999999Z0700", // ISO8601 with nanoseconds
+	if !strings.Contains(candidate, ":") && !strings.Contains(candidate, "-") {
+		// The blow formats either have a : or a - in them. If the string contains neither it cannot be a time string
+		return time.Now(), false
+	}
+
+	timeFormats := [...]timeFormat{
+		{time.ANSIC, len(time.ANSIC) - 1, len(time.ANSIC)},
+		{time.UnixDate, len(time.UnixDate) - 1, len(time.ANSIC)},
+		{time.RubyDate, len(time.RubyDate), len(time.RubyDate)},
+		{time.Kitchen, len(time.Kitchen), len(time.Kitchen) + 1},
+		{time.RFC3339, len(time.RFC3339), len(time.RFC3339)},
+		{time.RFC3339Nano, len(time.RFC3339Nano), len(time.RFC3339Nano)},
+		{"2006-01-02", 10, 10},                         // RFC 3339
+		{"2006-01-02 15:04", 16, 16},                   // RFC 3339 with minutes
+		{"2006-01-02 15:04:05", 19, 19},                // RFC 3339 with seconds
+		{"2006-01-02 15:04:05-07:00", 25, 25},          // RFC 3339 with seconds and timezone
+		{"2006-01-02T15Z0700", 18, 18},                 // ISO8601 with hour
+		{"2006-01-02T15:04Z0700", 21, 21},              // ISO8601 with minutes
+		{"2006-01-02T15:04:05Z0700", 24, 24},           // ISO8601 with seconds
+		{"2006-01-02T15:04:05.999999999Z0700", 34, 34}, // ISO8601 with nanoseconds
 	}
 
 	for _, format := range timeFormats {
-
-		ret, found = tryParseExactTime(candidate, format)
+		// Avoid trying to parse formats it could not be to reduce allocation of time parse errors
+		if len(candidate) < format.minLength || len(candidate) > format.maxLength {
+			continue
+		}
+		ret, found = tryParseExactTime(candidate, format.format)
 		if found {
 			return ret, true
 		}