--- a/tokenStream.go
+++ b/tokenStream.go
@@ -1,30 +1,43 @@
 package govaluate
 
+import "sync"
+
 type tokenStream struct {
 	tokens      []ExpressionToken
 	index       int
 	tokenLength int
 }
 
+var tokenStreamPool = sync.Pool{
+	New: func() interface{} {
+		return new(tokenStream)
+	},
+}
+
 func newTokenStream(tokens []ExpressionToken) *tokenStream {
-	ret := new(tokenStream)
+	ret := tokenStreamPool.Get().(*tokenStream)
 	ret.tokens = tokens
+	ret.index = 0
 	ret.tokenLength = len(tokens)
 	return ret
 }
 
-func (this *tokenStream) rewind() {
-	this.index -= 1
+func (t *tokenStream) rewind() {
+	t.index -= 1
 }
 
-func (this *tokenStream) next() ExpressionToken {
-	token := this.tokens[this.index]
+func (t *tokenStream) next() ExpressionToken {
+	token := t.tokens[t.index]
 
-	this.index += 1
+	t.index += 1
 	return token
 }
 
-func (this tokenStream) hasNext() bool {
+func (t tokenStream) hasNext() bool {
 
-	return this.index < this.tokenLength
+	return t.index < t.tokenLength
+}
+
+func (t *tokenStream) close() {
+	tokenStreamPool.Put(t)
 }