@see-sharp-dev 2025-05-13
@hsluoyz @pboyd04 this change started to break our tests (that run in parallel with `t.Parallel()`) because of the detected data races.
```
WARNING: DATA RACE
Read at 0x00c00003c170 by goroutine 28:
  github.com/casbin/govaluate.parseTokens()
      /xxx/vendor/github.com/casbin/govaluate/parsing.go:60 +0x548
  github.com/casbin/govaluate.NewEvaluableExpressionWithFunctions()
      /xxx/vendor/github.com/casbin/govaluate/EvaluableExpression.go:100 +0xb4
  github.com/casbin/casbin/v2.(*Enforcer).getAndStoreMatcherExpression()
      /xxx/vendor/github.com/casbin/casbin/v2/enforcer.go:838 +0x90
  github.com/casbin/casbin/v2.(*Enforcer).enforce()
      /xxx/vendor/github.com/casbin/casbin/v2/enforcer.go:696 +0xd10
  github.com/casbin/casbin/v2.(*Enforcer).Enforce()
      /xxx/vendor/github.com/casbin/casbin/v2/enforcer.go:849 +0x230
```