@see-sharp-dev 2025-05-13
@hsluoyz I'm not really familiar with the repo and code. we use casbin and after updating indirect modules our casbin tests started to fail.

an example of what I meant is [here](https://github.com/see-sharp-dev/govaluate/pull/1)