@pboyd04 2025-05-13
I can add some locking to prevent the "race". it shouldn't be a real problem since all that should be able to happen here is that one of the parallel govalute parses overwrites the sample size from another. This shouldn't have any impacts outside the library. But yeah I can make it cleaner.