@pboyd04 2025-05-13
It's trivial to fix. All you need is a mutex around the samples array. I'll do it when I get some time. Should be this evening or next.