@pboyd04 2025-05-13
Fixed in https://github.com/casbin/govaluate/pull/15